Clarify verify and sif commands are for local images
Note that verify and sif commands require a local image and cannot run
against library URIs directly.

While we are here, move the remote login from the verify instructions
to the signing instructions. You don't need a login to retrieve keys
for verification... and people may have a need to verify, but not
sign.

Signed-off-by: Dave Dykstra <2129743+DrDaveD@users.noreply.github.com>